requirements.txt
django-ninja==1.1.0
Django==5.0.4
psycopg2-binary==2.9.9
ruff==0.4.3 # for format
django-ninja-extra
django-ninja-jwt